# openapi.yaml
openapi: 3.0.0
info:
  title: ServiceDesk API
  version: 1.0.0
  description: |
    Enterprise Service Desk platform with multi-tenant support, comprehensive ticket management,
    knowledge base, analytics, and AI-powered features.

    ## Authentication
    This API uses JWT-based authentication with bearer tokens or secure HTTP-only cookies.
    All authenticated endpoints require a valid token obtained through the login endpoint.

    ## Multi-tenancy
    The API supports multi-tenant architecture. Tenant context is resolved via:
    - Subdomain (tenant.servicedesk.com)
    - HTTP headers (x-tenant-id, x-tenant-slug)
    - Cookie (tenant-context)

    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Rate limits vary by endpoint type:
    - Auth endpoints: More restrictive
    - General API endpoints: Standard limits

    Rate limit information is included in response headers:
    - X-RateLimit-Limit
    - X-RateLimit-Remaining
    - X-RateLimit-Reset
  contact:
    name: ServiceDesk API Support
    email: support@servicedesk.com
  license:
    name: Proprietary
    url: https://servicedesk.com/license
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://{tenant}.servicedesk.com/api
    description: Production server (multi-tenant)
    variables:
      tenant:
        default: demo
        description: Tenant subdomain slug
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint

    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: HTTP-only secure cookie containing JWT token
...
  schemas:
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message description"
        code:
          type: string
          example: "ERROR_CODE"
...
  responses:
    UnauthorizedError:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Authentication required"
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Acesso negado - permissão insuficiente"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Erro interno do servidor"
paths:
...
  /analytics:
    get:
      tags:
        - Analytics
      summary: Get analytics data
      description: Retrieve analytics and metrics data (admin only)
      operationId: getAnalytics
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: period
          in: query
          description: Time period for analytics
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
            default: 30d
        - name: type
          in: query
          description: Type of analytics data
          schema:
            type: string
            enum: [overview, tickets, trends, users, performance, all]
            default: overview
      responses:
        '200':
          description: Analytics data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  period:
                    type: string
                    example: "30d"
                  analytics:
                    type: object
                    properties:
                      overview:
                        type: object
                        properties:
                          totalTickets:
                            type: integer
                            example: 450
                          openTickets:
                            type: integer
                            example: 125
                          closedTickets:
                            type: integer
                            example: 325
                          avgResolutionTime:
                            type: number
                            example: 24.5
                            description: Average resolution time in hours
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/ServerError'

  # ==========================================
  # NOTIFICATION ENDPOINTS
  # ==========================================
...